handle corrupt_gone in sync(), add corrupt ignored notification, move notificationhandler to its own module
